refactoring (ai/rsc): extract utils, type-safe create-suspensed-chunk
